// src/api/model/bracketGroupOverviewMatchDto.ts
/**
 * Generated by orval v7.3.0 🍺
 * Do not edit manually.
 * Play.ID API
 * Esports API for Play.ID
 * OpenAPI spec version: 1.0
 */
import type { BracketGroupOverviewMatchParticipantDto } from './bracketGroupOverviewMatchParticipantDto';

export interface BracketGroupOverviewMatchDto {
  /** 매치 ID */
  id: number;
  /** 매치 번호 */
  matchNumber: number;
  /**
   * 매치 이름
   * @nullable
   */
  name?: string | null;
  /** 참가자 목록 */
  participants?: BracketGroupOverviewMatchParticipantDto[];
  /** 이전 매치 ID 목록 */
  prevMatchIds?: number[];
  /** 라운드 ID */
  roundId: number;
  /**
   * 경기 일자 (YYYY-MM-DD)
   * @nullable
   */
  scheduledDate?: string | null;
  /**
   * 경기 시간 (HH:mm:ss)
   * @nullable
   */
  scheduledTime?: string | null;
}