// src/api/model/createBracketDtoFormat.ts
/**
 * Generated by orval v7.3.0 🍺
 * Do not edit manually.
 * Play.ID API
 * Esports API for Play.ID
 * OpenAPI spec version: 1.0
 */

/**
 * 대진 포맷
 */
export type CreateBracketDtoFormat = typeof CreateBracketDtoFormat[keyof typeof CreateBracketDtoFormat];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const CreateBracketDtoFormat = {
  SINGLE_ELIMINATION: 'SINGLE_ELIMINATION',
  FREE_FOR_ALL: 'FREE_FOR_ALL',
} as const;